Streamline work flow for updating and installing
The special considerations for fivethirtyeight data are handled by a new function, coursekata_repos(), which just adds the repo to the list of other repos. A default repo is also included to make the package easier to use from the command line via new method coursekata_install() (AKA coursekata_update()).