Fully implements `gf_model()`
Will plot:
- empty model
- regression model with one explanatory variable
- group model with one explanatory variable

Chains to previous plots as long as the data argument does not change.
Will inherit the previous `gformula` and try to use that for the model
formula if one is not given. Otherwise a new formula can be given to plot
a different model on top of the data.

#BB-430 #review